pyproject.toml
[project]
name = "obd2-diagnostic-pro"
version = "1.0.0"
description = "Professional OBD2 Automotive Diagnostic Tool with AI Analysis"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.51.0",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.78.0",
    "psycopg2-binary>=2.9.10",
    "pyserial>=3.5",
    "python-obd>=0.7.1",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
]
